Move streamBlobToFile out of utils into the CLI
